+++ habit/src/main/java/app/habit/domain/factory/EvaluationPromptFactory.java
@@ -1,5 +1,6 @@
-package app.habit.domain;
+package app.habit.domain.factory;
 
+import app.habit.domain.HabitFormingPhase;
 import app.habit.service.gpt.request.PromptMessage;
 import app.habit.service.gpt.request.PromptRole;
 import app.habit.service.gpt.request.RequestPrompt;
