+++ habit/src/main/java/app/habit/service/gpt/OpenAiService.java
@@ -2,18 +2,18 @@ package app.habit.service.gpt;
 
 import app.habit.domain.Answer;
 import app.habit.service.gpt.coach.EvaluationGptCoach;
-import app.habit.domain.EvaluationPromptFactory;
+import app.habit.domain.factory.EvaluationPromptFactory;
 import app.habit.service.gpt.coach.FeedbackGptCoach;
 import app.habit.domain.FeedbackModule;
-import app.habit.domain.FeedbackPromptFactory;
+import app.habit.domain.factory.FeedbackPromptFactory;
 import app.habit.domain.FeedbackSession;
 import app.habit.domain.HabitAssessmentManager;
 import app.habit.domain.HabitFormingPhase;
 import app.habit.service.gpt.coach.PhaseGptCoach;
 import app.habit.service.gpt.coach.PreQuestionGptCoach;
-import app.habit.domain.PromptFactory;
+import app.habit.domain.factory.PromptFactory;
 import app.habit.domain.Question;
-import app.habit.domain.SpecificPhasePreQuestionPromptFactory;
+import app.habit.domain.factory.SpecificPhasePreQuestionPromptFactory;
 import app.habit.domain.Subject;
 import app.habit.dto.HabitPreQuestionRs;
 import app.habit.dto.PhaseEvaluationRq;
