+++ habit/src/test/java/app/habit/domain/factory/FeedbackPromptFactoryTest.java
@@ -1,7 +1,10 @@
-package app.habit.domain;
+package app.habit.domain.factory;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import app.habit.domain.FeedbackModule;
+import app.habit.domain.FeedbackSession;
+import app.habit.domain.factory.FeedbackPromptFactory;
 import app.habit.service.gpt.request.PromptMessage;
 import app.habit.service.gpt.request.PromptRole;
 import app.habit.service.gpt.request.RequestPrompt;
